moving percent to a float

+++ db/migrate/20160924044608_create_licenses.rb
@@ -12,7 +12,7 @@ class CreateLicenses < ActiveRecord::Migration
       t.string :detail_action
       t.string :amount_action
       t.integer :amount
-      t.integer :percent
+      t.float :percent
       t.integer :units
       t.string :ccy
       t.string :recurring_type
